add fix for incorrect size on android phones
For whatever reason, without this on my Android device window.innerWidth is reported as 980
when in landscape orientation (while the actual width is 1280px).
This causes an incorrect frame, that can also not be zoomed.
This meta tag seems to fix zooming, and it does not seem to bother desktop devices.

Let me know if you know the proper fix.